+++ src/steps/fetch-hardware/index.test.ts
@@ -1,7 +1,8 @@
-import { createStepContext } from '../../../../test';
 import { Recording, setupRecording } from '@jupiterone/integration-sdk-testing';
 
-import { fetchHardwareAssets } from '../index';
+import { createStepContext } from '../../../test';
+import { fetchHardwareAssets } from './index';
+import { fetchAccount } from '../fetch-account';
 
 let recording: Recording;
 
@@ -9,11 +10,9 @@ afterEach(async () => {
   await recording.stop();
 });
 
-test.skip('should process hardware entities', async () => {
-  jest.setTimeout(10000);
-
+test('fetchHardwareAssets', async () => {
   recording = setupRecording({
-    name: 'snipeit_hardware',
+    name: 'fetchHardwareAssets',
     directory: __dirname,
     options: {
       recordFailedRequests: false,
@@ -26,11 +25,23 @@ test.skip('should process hardware entities', async () => {
   });
 
   const context = createStepContext();
+  await fetchAccount(context);
   await fetchHardwareAssets(context);
 
-  expect(context.jobState.collectedRelationships).toHaveLength(2651);
   expect(context.jobState.collectedRelationships).toEqual(
     expect.arrayContaining([
+      expect.objectContaining({
+        _key: expect.any(String),
+        _class: 'PROVIDES',
+        _type: 'snipeit_account_provides_service',
+        displayName: 'PROVIDES',
+      }),
+      expect.objectContaining({
+        _key: expect.any(String),
+        _class: 'MANAGES',
+        _type: 'snipeit_account_manages_location',
+        displayName: 'MANAGES',
+      }),
       expect.objectContaining({
         _key: expect.any(String),
         _class: 'MANAGES',
@@ -57,7 +68,7 @@ test.skip('should process hardware entities', async () => {
           sourceEntityKey: expect.any(String),
           targetFilterKeys: [['_class', 'id', 'locationId']],
           targetEntity: expect.objectContaining({
-            _class: 'Device',
+            _class: ['Device'],
             id: expect.any(String),
             locationId: expect.any(Number),
           }),